+++ database/migrations/2024_01_01_000002_create_services_table.php
@@ -10,7 +10,7 @@ return new class extends Migration {
         Schema::create("services", function (Blueprint $table) {
             $table->id();
             $table->string("name");
-            $table->decimal("price", 15, 2);
+            $table->unsignedInteger("price");
             $table->text("description")->nullable();
             $table->boolean("status")->default(true);
             $table->timestamps();
